Use size_t for generated loop indices (#371)
* Use size_t for generated loop indices

* Use ssize_t for generated loop indices

* Use ssize_t for local index variables

+++ tests/codegen_refs/add_chain.c
@@ -1,18 +1,19 @@
 #include <stdint.h>
+#include <sys/types.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_add_f32(const float a[2][3], const float b[2][3], float out[2][3]) {
-    for (int64_t i0 = 0; i0 < 2; ++i0) {
-        for (int64_t i1 = 0; i1 < 3; ++i1) {
+    for (ssize_t i0 = 0; i0 < 2; ++i0) {
+        for (ssize_t i1 = 0; i1 < 3; ++i1) {
             out[i0][i1] = ref_scalar_f32_add(a[i0][i1], b[i0][i1]);
         }
     }
 }
 
 void node2_add_f32(const float a[2][3], const float b[2][3], float out[2][3]) {
-    for (int64_t i0 = 0; i0 < 2; ++i0) {
-        for (int64_t i1 = 0; i1 < 3; ++i1) {
+    for (ssize_t i0 = 0; i0 < 2; ++i0) {
+        for (ssize_t i1 = 0; i1 < 3; ++i1) {
             out[i0][i1] = ref_scalar_f32_add(a[i0][i1], b[i0][i1]);
         }
     }
