+++ tests/codegen_refs/argmax_dim.c
@@ -1,12 +1,13 @@
 #include <stdint.h>
+#include <sys/types.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_argmax_f32(const float a[2][3], int64_t out[2]) {
-    for (int64_t i0 = 0; i0 < 2; ++i0) {
+    for (ssize_t i0 = 0; i0 < 2; ++i0) {
         float best_value = a[i0][0];
-        int64_t best_index = 0;
-        for (int64_t r1 = 1; r1 < 3; ++r1) {
+        ssize_t best_index = 0;
+        for (ssize_t r1 = 1; r1 < 3; ++r1) {
             float value = a[i0][r1];
             if (value > best_value) {
                 best_value = value;
