+++ tests/codegen_refs/sum_global.c
@@ -1,11 +1,12 @@
 #include <stdint.h>
+#include <sys/types.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_sum_f32(const float a[2][3], float out[1]) {
     float acc = 0.0f;
-    for (int64_t r0 = 0; r0 < 2; ++r0) {
-        for (int64_t r1 = 0; r1 < 3; ++r1) {
+    for (ssize_t r0 = 0; r0 < 2; ++r0) {
+        for (ssize_t r1 = 0; r1 < 3; ++r1) {
             acc += a[r0][r1];
         }
     }
